# Conflicts:
#	.github/workflows/changelog.yml

+++ src/map/items/item.cpp
@@ -167,7 +167,7 @@ uint8 CItem::getAHCat() const
  *                                                                       *
  ************************************************************************/
 
-void CItem::setType(uint8 type)
+void CItem::setType(uint16 type)
 {
     m_type |= type;
 }
